backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.0.0
python-dotenv>=1.0.0
strands-agents>=1.10.0
requests
pytz
python-dateutil>=2.8.0
pandas>=1.5.0
urllib3>=1.26.0